Build: Allow skipping libxcb compile with env vars
Commit 0faf13e20d26446defe92c9b68f883892bbf928f allows users to not have
to use bundled libxcb at runtime but they still have to build it.

With this patch it's possible to pass
RADIUM_BUILD_LIBXCB=0 to make packages and
RADIUM_INSTALL_LIBXCB=0 to install.sh

to build without libxcb as well.

+++ bin/packages/build.sh
@@ -207,7 +207,7 @@ then
     cd ../
 fi
 
-if [[ $RADIUM_QT_VERSION == 5 ]]
+if [[ $RADIUM_QT_VERSION == 5 && $RADIUM_BUILD_LIBXCB != 0 ]]
 then
 
     rm -fr xcb-proto-1.13/
